updpkg(main/gnupg): update to 2.5.5
This version of gnupg does not ship its own yat2m (used for doc
generation) anymore. yat2m is provided by gpgrt-tools in ubuntu, but
the version shipped in the ubuntu 24.04 (that we currently use) is too
old.

To solve this we download libgpg-error sources and host-build the
latest version for gnupg.

Co-authored-by: Yaksh Bariya <yakshbari4@gmail.com>